+++ src/server/saved_tracker.rs
@@ -1,14 +1,14 @@
-use std::collections::HashMap;
+use std::{collections::HashMap, fs::create_dir_all, path::PathBuf};
 
 use anyhow::{Context, Result};
-use rspotify::prelude::Id;
-use rspotify::Config;
 use rspotify::{
-    clients::OAuthClient, model::TrackId, scopes, AuthCodeSpotify, Credentials,
-    OAuth,
+    clients::OAuthClient, model::TrackId, prelude::Id, scopes, AuthCodeSpotify,
+    Config, Credentials, OAuth, DEFAULT_CACHE_PATH,
 };
 use tokio::sync::mpsc::Sender;
 use tracing::{debug, info};
+
+use crate::shared::config::{resolve_home_path, DEFAULT_CONFIG_FOLDER};
 
 use super::grpc::api::ChangeEvent;
 
@@ -44,10 +44,17 @@ impl SavedTracker {
         let creds = Credentials::from_env().context(
             "Could not find RSPOTIFY_CLIENT_(ID|SECRET) environment variables",
         )?;
+        let cache_folder =
+            resolve_home_path(PathBuf::from(DEFAULT_CONFIG_FOLDER))?;
+        create_dir_all(cache_folder.clone()).with_context(|| {
+            format!("Could not create cache folder {cache_folder:?}")
+        })?;
+        let cache_path = cache_folder.join(DEFAULT_CACHE_PATH);
+        // error!("{cache_path}");
         let config = Config {
             token_cached: true,
             token_refreshing: true,
-            cache_path: "/home/andreas/.spotify_token_cache.json".into(),
+            cache_path,
             ..Default::default()
         };
         let mut spotify = AuthCodeSpotify::with_config(creds, oauth, config);